+++ CHANGELOG.md
@@ -7,11 +7,14 @@
 - Requires `innmind/foundation:~1.9`
 - Requires `innmind/di:~3.0`
 - `Innmind\Framework\Application::route()` callable must now return a `Innmind\Router\Component`
+- `Innmind\Framework\Application::route()` callable first parameter now is a `Innmint\Router\Pipe`
+- `Innmind\Framework\Application::route()` first parameter must now be expressed via a component inside the callable
 
 ### Removed
 
 - The ability to use `string`s to reference services
 - `Innmind\Framework\Http\Service`
+- `Innmind\Framework\Http\To`
 
 ### Fixed